fixing minor bugs in ppo lirl

+++ agents/ppo_lirl.py
@@ -297,7 +297,7 @@ class PPO_Lirl(BaseAgent):
                 loss2 = torch.nn.MSELoss()(adv_dones.squeeze(), log_prob_act_batch[flt])
 
                 # just scaling the losses according to number of observations:
-                coef = flt.mean()
+                coef = flt.float().mean()
                 rew_loss = (1 - coef) * loss + coef * loss2
 
                 # detach grads for next_rew?
